% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.R
\name{CategoriesOnPath}
\alias{CategoriesOnPath}
\title{Title}
\usage{
CategoriesOnPath(TransfData, CellClass, PrinGraph, Projections, Path = "ask",
  Net = NULL, PathType = "Long.Linear", Circular = FALSE)
}
\arguments{
\item{TransfData}{}

\item{CellClass}{}

\item{PrinGraph}{}

\item{Projections}{}

\item{Path}{}

\item{Net}{}

\item{PathType}{}

\item{Circular}{}
}
\description{
Title
}
