% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.R
\name{CheckAndGetPath}
\alias{CheckAndGetPath}
\title{Title}
\usage{
CheckAndGetPath(Results, Data, Categories, Graph, Path, PathType, Circular)
}
\arguments{
\item{Results}{}

\item{Data}{}

\item{Categories}{}

\item{Graph}{}

\item{Path}{}
}
\description{
Title
}

