% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{CheckDirectionality}
\alias{CheckDirectionality}
\title{Title}
\usage{
CheckDirectionality(Data, Results, OrderClass, Depth = NULL)
}
\arguments{
\item{Data}{}

\item{Results}{}

\item{OrderClass}{}

\item{Depth}{}
}
\description{
Title
}

