% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principal_graphs.R
\name{CircleConfiguration}
\alias{CircleConfiguration}
\title{Define the grammar to produde a circular principal graph}
\usage{
CircleConfiguration()
}
\value{
A parameter set that can be used to construct a circular principal graph
}
\description{
This is an internal function and should not be accessed directly by the user.
}

