% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principal_graphs.R
\name{computeElasticPrincipalGraph}
\alias{computeElasticPrincipalGraph}
\title{Compute the elastic principal graph}
\usage{
computeElasticPrincipalGraph(Data, NumNodes, NodesPositions = NULL,
  Edges = NULL, Method = NULL, NodeStep = NULL, ReduceMethod = "none",
  ReduceSize = 3, ...)
}
\arguments{
\item{Data}{The datapoints to be used. Rows are expected to indicate the points and columes their dimensions}

\item{NumNodes}{The number of nodes to be used to compute the principal graph}

\item{Method}{The parametriation associated with a specific graph grammar.}

\item{ReduceMethod}{An optional string indicating if dimensionality reduction via PCA need to per performed}

\item{ReduceSize}{Th number of dimensions to retain in case of dimensionality reduction}

\item{...}{Additional parameters to be passed to the internal function thqt computes the principal graph}
}
\value{
A list describing the principal elastic graph
}
\description{
Compute the elastic principal graph
}

