% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{ConstructGraph}
\alias{ConstructGraph}
\title{Title}
\usage{
ConstructGraph(Results, DirectionMat = NULL, Thr = 0.05)
}
\arguments{
\item{Results}{}

\item{DirectionMat}{}

\item{Thr}{}
}
\description{
Title
}
