% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{ConvergeOnGenes}
\alias{ConvergeOnGenes}
\title{Title}
\usage{
ConvergeOnGenes(ExpData, StartGeneSet, Mode = "VarALL", DistillThr = 0.5,
  ExtMode = 1, StopCrit = 0.99, MaxRounds = 25, FastReduce = FALSE,
  OutThr = 3, nNodes = 25, VarThr = 0.99, Categories = NULL,
  UpdateG0Cells = TRUE, GraphType = "Circle", IgnoreTail = FALSE,
  PlanVarLimit = 0.9, PlanVarLimitIC = 0.95, ForceLasso = FALSE,
  InitStructNodes = 10, MinBranDiff = 2, Log = TRUE, Filter = TRUE,
  MinProlCells = 25, DipPVThr = 1e-04, PCACenter = FALSE,
  PCAProjCenter = TRUE, PlotIntermediate = FALSE, StartQuant = 0.5,
  QuantInc = 0.01)
}
\arguments{
\item{ExpData}{}

\item{StartGeneSet}{}

\item{Mode}{}

\item{DistillThr}{}

\item{ExtMode}{}

\item{StopCrit}{}

\item{OutThr}{}

\item{nNodes}{}

\item{VarThr}{}

\item{Categories}{}

\item{GraphType}{}

\item{PlanVarLimit}{}

\item{PlanVarLimitIC}{}

\item{ForceLasso}{}

\item{MinBranDiff}{}

\item{Log}{}

\item{Filter}{}

\item{MinProlCells}{}

\item{DipPVThr}{}

\item{PCACenter}{}

\item{PCAProjCenter}{}

\item{PlotIntermediate}{}

\item{LassoCircInit}{}
}
\description{
Title
}
