% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{ConvertNames}
\alias{ConvertNames}
\title{Conversion between gene naming convenctions and organisms}
\usage{
ConvertNames(SourceOrganism, TargetOrganism, Genes, SourceTypes = "Names",
  TargetTypes = "Names", HomologyLevel = 1)
}
\arguments{
\item{SourceOrganism}{Source organism (currently either "human" or "mouse")}

\item{TargetOrganism}{Target organism (currently either "human" or "mouse")}

\item{Genes}{Vector of strings containing gene names. NAs can be present and will be removed}

\item{SourceTypes}{Type of gene names used as input (currently either "Names" or "Ensembl"). Default is "Names".}

\item{TargetTypes}{Type of gene names to be returned (currently either "Names" or "Ensembl"). Default is "Names".}

\item{HomologyLevel}{minimal level of homology (1 by default)}
}
\description{
Conversion between gene naming convenctions and organisms
}
\examples{

GenNames <- c("CCND3", "CD151", "CD2BP2", "CD81", "CDC34", "CDC37", "CDC42BPB", "CDC42EP1","CDC42EP4")

ConvertNames("human", "mouse", SourceTypes = "Names", TargetTypes = "Names", GenNames)

MoN <- ConvertNames("human", "mouse", SourceTypes = "Names", TargetTypes = "Names", GenNames)
MoEn <- ConvertNames("human", "mouse", SourceTypes = "Names", TargetTypes = "Ensembl", GenNames)

ConvertNames("mouse", "human", SourceTypes = "Names", TargetTypes = "Names", MoN)
ConvertNames("mouse", "human", SourceTypes = "Ensembl", TargetTypes = "Ensembl", MoEn)


}

