% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principal_graphs.R
\name{CurveConfiguration}
\alias{CurveConfiguration}
\title{Define the grammar to produde a curve (a path) principal graph}
\usage{
CurveConfiguration()
}
\value{
A parameter set that can be used to construct a curve (a path) principal graph
}
\description{
This is an internal function and should not be accessed directly by the user.
}
