% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principal_graphs.R
\name{DefaultPrincipalTreeConfiguration}
\alias{DefaultPrincipalTreeConfiguration}
\title{Define the grammar to produde a tree principal graph}
\usage{
DefaultPrincipalTreeConfiguration()
}
\value{
A parameter set that can be used to construct a tree principal graph
}
\description{
This is an internal function and should not be accessed directly by the user.
}

