% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.R
\name{DistanceStatistics}
\alias{DistanceStatistics}
\title{Title}
\usage{
DistanceStatistics(ExpressionMatrix, NodeOnGenes, NodePos, CellPos)
}
\arguments{
\item{ExpressionMatrix}{}

\item{NodeOnGenes}{}

\item{NodePos}{}

\item{CellPos}{}
}
\description{
Title
}
