% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{DistillGene}
\alias{DistillGene}
\title{Title}
\usage{
DistillGene(BaseAnalysis, Mode = "VarPC", DistillThr = 1e-04,
  Topo = "lasso", FullExpression, LoesSpan = 0.1)
}
\arguments{
\item{BaseAnalysis}{}

\item{Mode}{}

\item{DistillThr}{}

\item{Topo}{}

\item{FullExpression}{}

\item{LoesSpan}{}
}
\description{
Title
}
