% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{Filter}
\alias{Filter}
\title{Title}
\usage{
Filter(ExpressionMatrix, GeneDetectedFilter, GeneCountFilter, MinCellExp,
  LogTranform, QuantNorm, GeneSet, Centering, Scaling)
}
\arguments{
\item{ExpressionMatrix}{}

\item{GeneDetectedFilter}{}

\item{GeneCountFilter}{}

\item{MinCellExp}{}

\item{LogTranform}{}

\item{QuantNorm}{}

\item{GeneSet}{}

\item{Centering}{}

\item{Scaling}{}
}
\description{
Title
}

