% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.R
\name{GeneExpressiononPath}
\alias{GeneExpressiononPath}
\title{Title}
\usage{
GeneExpressiononPath(ExpressionData, TransfData, CellClass = NULL, PrinGraph,
  Projections, Genes, Path = "ask", Net = NULL, PathType = "Long.Linear",
  Circular = FALSE)
}
\arguments{
\item{ExpressionData}{}

\item{CellClass}{}

\item{PrinGraph}{}

\item{Projections}{}

\item{Genes}{}

\item{Path}{}
}
\description{
Title
}

