% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{GetLongestPath}
\alias{GetLongestPath}
\title{Title}
\usage{
GetLongestPath(Net, Structure = "auto", Circular = TRUE)
}
\arguments{
\item{Net}{}

\item{Structure}{}

\item{Circular}{}
}
\description{
Title
}
