% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principal_graphs.R
\name{getTaxonMap}
\alias{getTaxonMap}
\title{Project a set of datapoints onto the nodes of principal graph}
\usage{
getTaxonMap(Results, Data, UseR = FALSE)
}
\arguments{
\item{Data}{A matrix containing a set of points with the correct number of dimensions. Each row represent a point}

\item{UseR}{A bolean indicating the use of R (TRUE) of the Java library to perform the projections}

\item{PrintGraph}{An elastic principal graph structure as returned from \code{link{computeElasticPrincipalGraph}}}
}
\value{
A list. Each elements of the list represent the vertex of the principal graph and contains a vector reporting
the cells (as row number) that are associted with that particular vertex. The vector equal to NA indicates that no
cells are associated with that vertex;
}
\description{
The function depends on a working JVM with the VDAOEngine.jar library in the libpath.
}

