% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadData.R
\name{LoadMatrixMarket}
\alias{LoadMatrixMarket}
\title{Title}
\usage{
LoadMatrixMarket(DataPath, GeneInfoPath = NULL, CellInfoPath = NULL,
  GeneCol = 1, CellCol = 1)
}
\arguments{
\item{DataPath}{}

\item{GeneInfoPath}{}

\item{CellInfoPath}{}

\item{GeneCol}{}

\item{CellCol}{}
}
\description{
Title
}
