% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principal_graphs.R
\name{makeGraph}
\alias{makeGraph}
\title{Produce a graph structure from the coputed elastic principal graph details}
\usage{
makeGraph(PrintGraph)
}
\arguments{
\item{PrintGraph}{An elastic principal graph structure as returned from \code{link{computeElasticPrincipalGraph}}}
}
\value{
A graph structure
}
\description{
The function depends on a working JVM with the VDAOEngine.jar library in the libpath.
This is an internal function and should not be accessed directly by the user
}

