% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering.R
\name{PCAOutLiers}
\alias{PCAOutLiers}
\title{Title}
\usage{
PCAOutLiers(DataMatrix, PCAMat = NULL, Dimensions, PCAMode = "base-svd",
  OutThr = 5)
}
\arguments{
\item{DataMatrix}{}

\item{PCAMat}{}

\item{Dimensions}{}

\item{PCAMode}{}

\item{OutThr}{}
}
\description{
Title
}
