% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotData2D}
\alias{plotData2D}
\title{Title}
\usage{
plotData2D(Data, PrintGraph, GroupsLab, ScaleFunction = sqrt,
  PlotProjections = FALSE, TaxonList = NULL, NodeSizeMult = 1,
  Col = NULL, CirCol = "black", ColLabels = NULL,
  ProjectionLines = NULL, OnEdgeProjections = NULL, LineCol = "black",
  IdCol = "blue", Main = "", Cex.Main = 0.7, Xlab = "PC1",
  Ylab = "PC2", Plot.ly = FALSE)
}
\arguments{
\item{Plot.ly}{}
}
\description{
Title
}
