% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotData3D}
\alias{plotData3D}
\title{Title}
\usage{
plotData3D(Data, PrintGraph, GroupsLab, ScaleFunction = sqrt,
  NodeSizeMult = 1, Col = NULL, CirCol = "black", LineCol = "black",
  IdCol = "blue", Main = "", Cex.Main = 0.7, PlotProjections = FALSE,
  ProjectionLines = NULL, TaxonList = NULL, Xlab = "PC1", Ylab = "PC2",
  Zlab = "PC3", DirectionMat = NULL, Thr = 0.05, Plot.ly = FALSE)
}
\arguments{
\item{Data}{}

\item{PrintGraph}{}

\item{GroupsLab}{}

\item{ScaleFunction}{}

\item{NodeSizeMult}{}

\item{Col}{}

\item{CirCol}{}

\item{LineCol}{}

\item{IdCol}{}

\item{Main}{}

\item{Cex.Main}{}

\item{PlotProjections}{}

\item{ProjectionLines}{}

\item{TaxonList}{}

\item{Xlab}{}

\item{Ylab}{}

\item{Zlab}{}

\item{DirectionMat}{}

\item{Thr}{}

\item{Plot.ly}{}
}
\description{
Title
}

