% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{PlotOnStages}
\alias{PlotOnStages}
\title{Title}
\usage{
PlotOnStages(Structure, TaxonList, Categories, PrinGraph, Net, SelThr = NULL,
  ComputeOverlaps = TRUE, RotatioMatrix, PointProjections, ExpData,
  nGenes = 10)
}
\arguments{
\item{Structure}{}

\item{TaxonList}{}

\item{Categories}{}

\item{PrinGraph}{}

\item{Net}{}

\item{SelThr}{}

\item{ComputeOverlaps}{}

\item{RotatioMatrix}{}

\item{ExpData}{}
}
\description{
Title
}
