% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{PlotsAndDistill}
\alias{PlotsAndDistill}
\title{Title}
\usage{
PlotsAndDistill(GeneExprMat, StartSet, Categories, Topology = "Circle",
  IgnoreTail = FALSE, PlanVarLimit = 0.9, PlanVarLimitIC = 0.95,
  DistillThr = 0.7, Log = TRUE, StartQuant = 0.5, OutThr = 3,
  OutThrPCA = 3, Title = "", MinProlCells = 20, PCACenter = FALSE,
  PlotDebug = FALSE, Mode = "VarPC", ExtMode = 2, nNodes = 40,
  InitStructNodes = 20, DipPVThr = 1e-04, PCAFilter = TRUE,
  PCAProjCenter = TRUE, StopCrit = 0.95, Filter = TRUE, QuaThr = 0.5,
  CompareSet = list(), EstProlif = "MeanPerc")
}
\arguments{
\item{GeneExprMat}{}

\item{StartSet}{}

\item{Categories}{}

\item{Topology}{}

\item{IgnoreTail}{}

\item{PlanVarLimit}{}

\item{PlanVarLimitIC}{}

\item{DistillThr}{}

\item{Log}{}

\item{StartQuant}{}

\item{OutThr}{}

\item{OutThrPCA}{}

\item{Title}{}

\item{MinProlCells}{}

\item{PCACenter}{}

\item{PlotDebug}{}

\item{Mode}{}

\item{ExtMode}{}

\item{nNodes}{}

\item{InitStructNodes}{}

\item{DipPVThr}{}

\item{PCAFilter}{}

\item{PCAProjCenter}{}

\item{StopCrit}{}

\item{Filter}{}

\item{CompareSet}{}
}
\description{
Title
}
