% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{ProjectAndCompute}
\alias{ProjectAndCompute}
\title{Title}
\usage{
ProjectAndCompute(DataSet, GeneSet, OutThr, VarThr, nNodes, Log = TRUE,
  Categories = NULL, Filter = TRUE, GraphType = "Lasso",
  PlanVarLimit = 0.9, PlanVarLimitIC = NULL, MinBranDiff = 2,
  LassoCircInit = 8, ForceLasso = FALSE, DipPVThr = 0.001,
  MinProlCells = 25)
}
\arguments{
\item{DataSet}{}

\item{GeneSet}{}

\item{OutThr}{}

\item{VarThr}{}

\item{nNodes}{}

\item{Log}{}

\item{Categories}{}

\item{Filter}{}

\item{GraphType}{}

\item{PlanVarLimit}{}

\item{PlanVarLimitIC}{}

\item{MinBranDiff}{}

\item{LassoCircInit}{}

\item{ForceLasso}{}

\item{DipPVThr}{}

\item{MinProlCells}{}
}
\description{
Title
}
\examples{

}

