% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ProjectOnCircle}
\alias{ProjectOnCircle}
\title{Title}
\usage{
ProjectOnCircle(Points, Edges, Nodes, Categories, Title, ExpValues,
  PCACenter = TRUE)
}
\arguments{
\item{Points}{}

\item{Edges}{}

\item{Nodes}{}

\item{Categories}{}

\item{Title}{}

\item{ExpValues}{}

\item{PCACenter}{}
}
\description{
Title
}
