% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ProjectOnPrincipalGraph}
\alias{ProjectOnPrincipalGraph}
\title{Title}
\usage{
ProjectOnPrincipalGraph(Nodes, Edges, Points, UsedPoints = NULL,
  Categories = NULL, Title = "", PCACenter = TRUE, ShowFitted = TRUE)
}
\arguments{
\item{Nodes}{}

\item{Edges}{}

\item{Points}{}

\item{UsedPoints}{}

\item{Categories}{}

\item{Title}{}
}
\description{
Title
}
