% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{Rand.POrd.Cor}
\alias{Rand.POrd.Cor}
\title{Sample correlation with partial order}
\usage{
Rand.POrd.Cor(OrdVect, Vect1, Vect2, Round)
}
\arguments{
\item{OrdVect}{}

\item{Vect1}{}

\item{Vect2}{}

\item{Round}{}
}
\description{
Sample correlation with partial order
}

