% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principal_graphs.R
\name{RobustPrincipalTreeConfiguration}
\alias{RobustPrincipalTreeConfiguration}
\title{Define the grammar to produde a robust tree principal graph}
\usage{
RobustPrincipalTreeConfiguration()
}
\value{
A parameter set that can be used to construct a robust tree principal graph
}
\description{
This is an internal function and should not be accessed directly by the user
}
