% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.r
\name{SelectComputePCA}
\alias{SelectComputePCA}
\title{Compute the PCA for a given numeric matrix usig different methodologies}
\usage{
SelectComputePCA(DataMatrix, Components = NULL, Method = "base-svd", ...)
}
\arguments{
\item{DataMatrix}{The matrix used to compute PCA.}

\item{Components}{The number of principal components to be retained.}

\item{Method}{The method to be used to compute PCA. The current implementation accepts 'base-svd',
'base-cov', 'flashPCA-eigen', 'flashPCA-svd', and 'irlba-Lanczos'}
}
\value{
A list composed five elements rotation

Comp the rotation matrix

Center A vector containing the velue used to center the original matrix
}
\description{
Compute the PCA for a given numeric matrix usig different methodologies
}
\examples{
add(1, 1)
add(10, 1)
}
