% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principal_graphs.R
\name{TwoStagesRobustPrincipalTreeConfiguration}
\alias{TwoStagesRobustPrincipalTreeConfiguration}
\title{Define the grammar to produde a tree principal graph using a two stage approach.}
\usage{
TwoStagesRobustPrincipalTreeConfiguration()
}
\value{
A parameter set that can be used to construct a robust tree principal graph
}
\description{
This is an internal function and should not be accessed directly by the user.
This function is still experimental and not is not working yet.
}
