% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{DistillGene}
\alias{DistillGene}
\title{Filter Gene saccording to predefined properties}
\usage{
DistillGene(BaseAnalysis, Mode = "VarPC", DistillThr = 0.2, ExtMode = 1,
  Topo = "Lasso", IgnoreTail = FALSE, ZeroQuant = 0.75,
  FullExpression = NULL, Log = FALSE, LoesSpan = 0.1,
  MedianRatThr = 0.05, VarTestPVThr = 1e-05)
}
\arguments{
\item{BaseAnalysis}{a strure returned by Project and Compute}

\item{Mode}{scalar, a string indicating the algorithm to use. I can be
"VarPC" (Variance-based on genes the genes used to contruct the principal curve)
"PeakPC" (Peack-based on genes the genes used to contruct the principal curve)
"VarALL" (Variance-based on all genes)
"PeakALL" (Peack-based on all genes)}

\item{DistillThr}{numeric, a threshold used by the algorithm. It is a p-value for variance based selection
and a quantile for peack based selection}

\item{ExtMode}{}

\item{Topo}{string, a string describing the topology of the principal curve. It can be 'Lasso' or 'Circle'}

\item{IgnoreTail}{}

\item{ZeroQuant}{}

\item{FullExpression}{the matrix containing the full expression data (note that sample filtered during the construction of the principal graph will be removed)}

\item{Log}{}

\item{LoesSpan}{the span parameter to use with the loes function}

\item{MedianRatThr}{}

\item{VarTestPVThr}{}
}
\description{
Filter Gene saccording to predefined properties
}
