% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotData2D}
\alias{plotData2D}
\title{Title}
\usage{
plotData2D(Data, PrintGraph, GroupsLab, ScaleFunction = sqrt,
  PlotProjections = FALSE, TaxonList = NULL, NodeSizeMult = 1,
  Col = NULL, CirCol = "black", ColLabels = NULL,
  ProjectionLines = NULL, OnEdgeProjections = NULL, LineCol = "black",
  IdCol = "blue", Main = "", Cex.Main = 0.7, Xlab = "PC1",
  Ylab = "PC2", Plot.ly = FALSE)
}
\arguments{
\item{Data}{}

\item{PrintGraph}{}

\item{GroupsLab}{}

\item{ScaleFunction}{}

\item{NodeSizeMult}{}

\item{Col}{}

\item{CirCol}{}

\item{ColLabels}{}

\item{LineCol}{}

\item{IdCol}{}

\item{Main}{}

\item{Cex.Main}{}

\item{Xlab}{}

\item{Ylab}{}

\item{Plot.ly}{}
}
\description{
Title
}

