% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotPieNet}
\alias{plotPieNet}
\title{Title}
\usage{
plotPieNet(Results, Data, Categories, Graph = NULL, TaxonList = NULL,
  LayOut = "nicely", Main = "", ScaleFunction = sqrt, NodeSizeMult = 1,
  ColCat = NULL, PlotNet = TRUE, DirectionMat = NULL, Thr = 0.05,
  Arrow.size = 0.5, BiggestComponents = FALSE)
}
\arguments{
\item{Results}{}

\item{Data}{}

\item{Categories}{}

\item{Graph}{}

\item{TaxonList}{}

\item{LayOut}{}

\item{Main}{}

\item{ScaleFunction}{}

\item{NodeSizeMult}{}

\item{ColCat}{}

\item{DirectionMat}{}

\item{Thr}{}

\item{Arrow.size}{}

\item{BiggestComponents}{}
}
\description{
Title
}
