% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{ProjectAndCompute}
\alias{ProjectAndCompute}
\title{Title}
\usage{
ProjectAndCompute(DataSet, GeneSet = NULL, VarThr, nNodes, Log = TRUE,
  Categories = NULL, Filter = TRUE, OutThr = 5, PCAFilter = FALSE,
  OutThrPCA = 5, GraphType = "Lasso", PlanVarLimit = 0.9,
  PlanVarLimitIC = NULL, MinBranDiff = 2, InitStructNodes = 8,
  ForceLasso = FALSE, EstProlif = "MeanPerc", QuaThr = 0.5,
  NonG0Cell = NULL, DipPVThr = 0.001, MinProlCells = 25,
  PCACenter = TRUE, PCAProjCenter = FALSE, PlotDebug = FALSE,
  PlotIntermediate = FALSE)
}
\arguments{
\item{DataSet}{}

\item{GeneSet}{}

\item{VarThr}{}

\item{nNodes}{}

\item{Log}{}

\item{Categories}{}

\item{Filter}{}

\item{OutThr}{}

\item{GraphType}{}

\item{PlanVarLimit}{}

\item{PlanVarLimitIC}{}

\item{MinBranDiff}{}

\item{InitStructNodes}{}

\item{ForceLasso}{}

\item{DipPVThr}{}

\item{MinProlCells}{}
}
\description{
Title
}
\examples{

}
