% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{ProjectAndCompute}
\alias{ProjectAndCompute}
\title{Title}
\usage{
ProjectAndCompute(DataSet, GeneSet = NULL, OutThr, VarThr, nNodes,
  Log = TRUE, Categories = NULL, Filter = TRUE, GraphType = "Lasso",
  PlanVarLimit = 0.9, PlanVarLimitIC = NULL, MinBranDiff = 2,
  InitStructNodes = 8, ForceLasso = FALSE, EstProlif = "MeanPerc",
  QuaThr = 0.6, NonG0Cell = NULL, DipPVThr = 0.001, MinProlCells = 25,
  PCACenter = FALSE, PCAProjCenter = FALSE, PlotDebug = FALSE,
  PlotIntermediate = FALSE)
}
\arguments{
\item{DataSet}{}

\item{GeneSet}{}

\item{OutThr}{}

\item{VarThr}{}

\item{nNodes}{}

\item{Log}{}

\item{Categories}{}

\item{Filter}{}

\item{GraphType}{}

\item{PlanVarLimit}{}

\item{PlanVarLimitIC}{}

\item{MinBranDiff}{}

\item{InitStructNodes}{}

\item{ForceLasso}{}

\item{DipPVThr}{}

\item{MinProlCells}{}
}
\description{
Title
}
\examples{

}
