% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{ProjectAndExpand}
\alias{ProjectAndExpand}
\title{Title}
\usage{
ProjectAndExpand(GeneExprMat, StartSet, Categories, Topology = "Circle",
  DistillThr = 0.6, IgnoreTail = TRUE, Log = TRUE, StartQuant = 0.25,
  Title = "Buettner et al", PlanVarLimit = 0.85, PlanVarLimitIC = 0.9,
  KeepOriginal = TRUE, PCACenter = FALSE, PlotDebug = FALSE,
  Mode = "VarPC", ExtMode = 3, MaxRounds = 15, StopCrit = 0.95,
  ExpQuant = 0.01, StopMode = 1, Parallel = TRUE, nCores = NULL,
  ClusType = "PSOCK")
}
\arguments{
\item{GeneExprMat}{}

\item{StartSet}{}

\item{Categories}{}

\item{Topology}{}

\item{DistillThr}{}

\item{IgnoreTail}{}

\item{Log}{}

\item{StartQuant}{}

\item{Title}{}

\item{PlanVarLimit}{}

\item{PlanVarLimitIC}{}

\item{KeepOriginal}{}

\item{PCACenter}{}

\item{PlotDebug}{}

\item{Mode}{}

\item{ExtMode}{}

\item{MaxRounds}{}

\item{StopCrit}{}

\item{ExpQuant}{}

\item{StopMode}{}

\item{Parallel}{}

\item{nCores}{}

\item{ClusType}{}
}
\description{
Title
}
