% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ProjectOnCircle}
\alias{ProjectOnCircle}
\title{Title}
\usage{
ProjectOnCircle(Points, Edges, Nodes, Categories, Title, ExpValues,
  PCACenter = TRUE, PlotPC3 = FALSE, PointAlpha = 0.3)
}
\arguments{
\item{PCACenter}{}
}
\description{
Title
}
