% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projections.R
\name{projectPoints}
\alias{projectPoints}
\title{Project Points on the edges of the graph}
\usage{
projectPoints(Results, Data, TaxonList = NULL, UseR = TRUE,
  method = "PCALin", Dims = NULL, Debug = FALSE)
}
\arguments{
\item{Results}{}

\item{Data}{}

\item{TaxonList}{}

\item{UseR}{}

\item{method}{}
}
\description{
Project Points on the edges of the graph
}
