% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipelines.R
\name{StudyCellCycles}
\alias{StudyCellCycles}
\title{Title}
\usage{
StudyCellCycles(ExpressionMatrix, Grouping, GeneSet = NULL,
  StageAssociation = NULL, TopQ = 0.9, LowQ = 0.1, NodePower = 2,
  PathOpt = "switch", GeneOpt = 10, GeneDetectedFilter = 2.5,
  GeneCountFilter = 2.5, MinCellExp = 1, VarFilter = 0,
  LogTranform = TRUE, Centering = FALSE, Scaling = FALSE, nDim = NULL,
  nPoints = 20, Data.Return = FALSE, GeneToPlot = 10, ThrNumb = NULL,
  Interactive = TRUE)
}
\arguments{
\item{ExpressionMatrix}{}

\item{Grouping}{}

\item{GeneSet}{}

\item{StageAssociation}{}

\item{TopQ}{}

\item{LowQ}{}

\item{PathOpt}{}

\item{GeneOpt}{}

\item{GeneDetectedFilter}{}

\item{GeneCountFilter}{}

\item{MinCellExp}{}

\item{VarFilter}{}

\item{LogTranform}{}

\item{Centering}{}

\item{Scaling}{}

\item{nDim}{}

\item{nPoints}{}

\item{Data.Return}{}

\item{GeneToPlot}{}

\item{AggressiveStaging}{}
}
\description{
Title
}

